oci-systemd-hook: use $container_uuid= when the container has it
we were using runc id for the machine-id which is wrong, as the runc
id is not an uuid.  Change it to use the $container_uuid environment
variable when defined.  As an effect, we mount the correct
/var/log/journal/%s directory in the container.

This means that containers using systemd must define a container_uuid=
environment variable.  To simplify this task for system containers, we
can automatically generate a new variable, so that the environment
block will look like:

        "env": [
...
            "container_uuid=$CONTAINER_UUID"
...
        ],

Signed-off-by: Giuseppe Scrivano <gscrivan@redhat.com>