Don't mount tmpfs on /var/log since it can cover other directories
We were mounting a tmpfs on /var/log so that we could handle read-only
containers, the problem with this is that it will cover over other
content in /var/log that might be needed by other tools.

This patch attempt to mount journal on /run/journal if /var/log/journal
fails. Systemd should attempt to write journal information to this
directory if /var/log is not writable.