improve error message if no 'bundlePath' in stateData
systemdhook exits through the same error handling code path if it fails
to open the config file or if the config file name is not specified via
'bundlePath' in stateData. This patch ensures that the latter error is
reported explicitly:

  Failed to open config file: bundlePath not found in state

Also, call snprintf() to fill the 'config_file_name' buffer instead of
sprintf() for better protection against buffer overflow.