introduce the getJSONstring() function
This is a generic function that can, for example, be called to read
the state JSON data from standard input or the config JSON data from
the location pointed to by 'bundlePath'. Buffers to store the data
in memory are allocated dynamically to remove previously existing
limitations in terms of the state data size and config data size
that systemdhook is able to parse.