Unmount mounts from inside container.
Unmount certain mounts as specified in /etc/oci-umount.conf from inside
container when container starts.

This patch fixes various issues, improves error handling.

Also checks if mount is actually mounted inside container or not. If not,
it skips it (instead of erroring out). We parse /proc/self/mountinfo to
figure out if a path is a mountpoint or not inside container.

Signed-off-by: Vivek Goyal <vgoyal@redhat.com>