Parse config mounts only for prestart hook
Right now parsing logic is in main(). That means this code gets executed
every time hook is called. We need to parse this only if it is prestart()
hook.

So create another function parseBundle() and move most of the parsing code
in there and call it only if this is prestart hook.

As we moved yajl node config_data in a funciton, it gets freed when
parseBundle() finishes. So can't store pointers pointing to strings in
that tree. Hence I am doing strdup() and storing a copy. Also freeing
config_mounts using gcc cleanup semantics.

Signed-off-by: Vivek Goyal <vgoyal@redhat.com>