Allow container definitions where rootfs is not an absolute path
Patch copied from https://github.com/projectatomic/oci-systemd-hook/pull/67

From Jason Wessel

In earlier versions of the runc frame work the rootfs path was passed
as a key with the initial json that was passed on the stdin and it was
automatically computed to be an absolute path.

This translation to an absolute path must be done in the hook based on
the bundlePath. This allows the config.json to be relocated by the
container hosting system storage without modifying the config.json.

Signed-off-by: Daniel J Walsh <dwalsh@redhat.com>