Move is_mounted() check inside unmount()
This check only needs to be done for the case when we are not mounting
submounts. For the case of submounts, path does not have to be a mount
point. So move this check inside unmount() and do this check only if
this is not the case of submounts of a path.

Signed-off-by: Vivek Goyal <vgoyal@redhat.com>