Link to libglib, use it in one place to start
libglib is a hard requirement of other projects on the host, such
as libostree.  Let's make use of it, since it has a lot of useful
functionality for C programs.

Specifically, it aborts on OOM by default (because in practice anything that
isn't the kernel/pid 1 isn't going to be able to continue), so we can just use
`g_strdup()`.

(Note that it's required nowadays that `g_malloc() = malloc()`, so it's
 OK that we're passing this to `free()`.)