Look in hooks environment for current state
We have changed CRI-O and podman to pass down the stage
as an environment variable "stage".  We also now allow
admins to specify which arguments can be passed into the hooks,
this patch causes oci-umount to look for the environment
variable and fall back to the old behaviour so it will continue
to work with Docker.

Signed-off-by: Daniel J Walsh <dwalsh@redhat.com>