Get hook stage from stdin's 'status' value
Instead of using CRI-O-specific arguments, environment variables, or
PID positive-ness, use the state's REQUIRED 'status' property [1] to
ensure we're being called on a created container (and not a running
container or a stopped container).  Also, do not error if 'pid' is
unset (it isn't REQUIRED for 'stopped' on Linux [2]) in general, but
do error if pid is not set positive for 'created' containers (where
it's REQUIRED on Linux [2]).

[1]: https://github.com/opencontainers/runtime-spec/blame/v1.0.1/runtime.md#L16
[2]: https://github.com/opencontainers/runtime-spec/blame/v1.0.1/runtime.md#L25